+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @nextui-org/popover
 
+## 0.0.0-dev-v2-20230407155017
+
+### Patch Changes
+
+- Popover a11y improved, titleProps passed by render props, form auto focus added
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230407155017
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230407155017
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230407155017
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230407155017
+  - @nextui-org/button@0.0.0-dev-v2-20230407155017
+  - @nextui-org/system@0.0.0-dev-v2-20230407155017
+  - @nextui-org/theme@0.0.0-dev-v2-20230407155017
+
+## 0.0.0-dev-v2-20230407131834
+
+### Patch Changes
+
+- Popover trigger scale down when is open
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230407131834
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230407131834
+  - @nextui-org/aria-utils@0.0.0-dev-v2-20230407131834
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230407131834
+  - @nextui-org/button@0.0.0-dev-v2-20230407131834
+  - @nextui-org/system@0.0.0-dev-v2-20230407131834
+  - @nextui-org/theme@0.0.0-dev-v2-20230407131834
+
 ## 0.0.0-dev-v2-20230407125356
 
 ### Patch Changes