+++ packages/core/react/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @nextui-org/react
 
+## 0.0.0-dev-v2-20230407155017
+
+### Patch Changes
+
+- Popover a11y improved, titleProps passed by render props, form auto focus added
+- Updated dependencies
+  - @nextui-org/pagination@0.0.0-dev-v2-20230407155017
+  - @nextui-org/accordion@0.0.0-dev-v2-20230407155017
+  - @nextui-org/checkbox@0.0.0-dev-v2-20230407155017
+  - @nextui-org/progress@0.0.0-dev-v2-20230407155017
+  - @nextui-org/popover@0.0.0-dev-v2-20230407155017
+  - @nextui-org/snippet@0.0.0-dev-v2-20230407155017
+  - @nextui-org/spinner@0.0.0-dev-v2-20230407155017
+  - @nextui-org/tooltip@0.0.0-dev-v2-20230407155017
+  - @nextui-org/avatar@0.0.0-dev-v2-20230407155017
+  - @nextui-org/button@0.0.0-dev-v2-20230407155017
+  - @nextui-org/switch@0.0.0-dev-v2-20230407155017
+  - @nextui-org/badge@0.0.0-dev-v2-20230407155017
+  - @nextui-org/input@0.0.0-dev-v2-20230407155017
+  - @nextui-org/radio@0.0.0-dev-v2-20230407155017
+  - @nextui-org/card@0.0.0-dev-v2-20230407155017
+  - @nextui-org/chip@0.0.0-dev-v2-20230407155017
+  - @nextui-org/code@0.0.0-dev-v2-20230407155017
+  - @nextui-org/drip@0.0.0-dev-v2-20230407155017
+  - @nextui-org/link@0.0.0-dev-v2-20230407155017
+  - @nextui-org/user@0.0.0-dev-v2-20230407155017
+  - @nextui-org/system@0.0.0-dev-v2-20230407155017
+  - @nextui-org/theme@0.0.0-dev-v2-20230407155017
+
+## 0.0.0-dev-v2-20230407131834
+
+### Patch Changes
+
+- Popover trigger scale down when is open
+- Updated dependencies
+  - @nextui-org/pagination@0.0.0-dev-v2-20230407131834
+  - @nextui-org/accordion@0.0.0-dev-v2-20230407131834
+  - @nextui-org/checkbox@0.0.0-dev-v2-20230407131834
+  - @nextui-org/progress@0.0.0-dev-v2-20230407131834
+  - @nextui-org/popover@0.0.0-dev-v2-20230407131834
+  - @nextui-org/snippet@0.0.0-dev-v2-20230407131834
+  - @nextui-org/spinner@0.0.0-dev-v2-20230407131834
+  - @nextui-org/tooltip@0.0.0-dev-v2-20230407131834
+  - @nextui-org/avatar@0.0.0-dev-v2-20230407131834
+  - @nextui-org/button@0.0.0-dev-v2-20230407131834
+  - @nextui-org/switch@0.0.0-dev-v2-20230407131834
+  - @nextui-org/badge@0.0.0-dev-v2-20230407131834
+  - @nextui-org/input@0.0.0-dev-v2-20230407131834
+  - @nextui-org/radio@0.0.0-dev-v2-20230407131834
+  - @nextui-org/card@0.0.0-dev-v2-20230407131834
+  - @nextui-org/chip@0.0.0-dev-v2-20230407131834
+  - @nextui-org/code@0.0.0-dev-v2-20230407131834
+  - @nextui-org/drip@0.0.0-dev-v2-20230407131834
+  - @nextui-org/link@0.0.0-dev-v2-20230407131834
+  - @nextui-org/user@0.0.0-dev-v2-20230407131834
+  - @nextui-org/system@0.0.0-dev-v2-20230407131834
+  - @nextui-org/theme@0.0.0-dev-v2-20230407131834
+
 ## 0.0.0-dev-v2-20230407125356
 
 ### Patch Changes