+++ packages/hooks/use-aria-field/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @nextui-org/use-aria-field
 
+## 0.0.0-dev-v2-20230407155017
+
+### Patch Changes
+
+- Popover a11y improved, titleProps passed by render props, form auto focus added
+- Updated dependencies
+  - @nextui-org/use-aria-slot-id@0.0.0-dev-v2-20230407155017
+  - @nextui-org/use-aria-label@0.0.0-dev-v2-20230407155017
+
+## 0.0.0-dev-v2-20230407131834
+
+### Patch Changes
+
+- Popover trigger scale down when is open
+- Updated dependencies
+  - @nextui-org/use-aria-slot-id@0.0.0-dev-v2-20230407131834
+  - @nextui-org/use-aria-label@0.0.0-dev-v2-20230407131834
+
 ## 0.0.0-dev-v2-20230407125356
 
 ### Patch Changes